+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Completely new interface
 - Improved handling of sub-queries and composition
-- Functional based criteria builders
+- Functional and string based criteria builders
 - Removal of all static methods
 - Requires PHP 7.1 or better